remove tire_helper
